Use dedicated TRACE class for traces that aren't special
dyn_cast<TRACK*> has the "interesting" feature that it will return NULL for
subclasses of TRACK, such as VIA and SEGZONE. To get the same behaviour
from dynamic_cast<>, we need to use a separate subclass.

Ideally, we'd mark TRACK::ClassOf as deleted to avoid mapping it to
BOARD_CONNECTED_ITEM::ClassOf, but that breaks SWIG (which generates a stub
for deleted functions).