Pcbnew: Prevent footprint jumping when starting to drag it
CHANGED: When dragging a footprint, if the mouse isn't directly over the
footprint anchor when 'D' is pressed, the footprint center will jump to
the mouse position when dragging begins.  This MR fixes the jumping
problem.  It also respects user preference for mouse warping, so that
dragging behaves like moving and warps the mouse as desired.

Fixes https://gitlab.com/kicad/code/kicad/issues/6813