Modifying DateAndTime() implementation for customization by user/programmer.
ADDED: DateAndTime() in kicad_string.h can return date with or without time.
Added parameters with default values so that current usage of this function is not modified.

ADDED: a panel in settings window with a checkbox to select time with date.
To be used in later version.

See : https://gitlab.com/kicad/code/kicad/-/issues/6121