cmake build system: CMAKE_MODULE_PATH use corrected
https://cmake.org/cmake/help/latest/variable/CMAKE_MODULE_PATH.html defines
that CMAKE_MODULE_PATH is supposed to be a semicolon-separated list of
paths to cmake module files. In Kicad this variable is used throughout as
a path description to CMakeModules directory.
Any third party module that uses CMAKE_MODULE_PATH as intended by Kitware
will break the build system.
pybind11 actually does append a path to CMAKE_MODULE_PATH but it didn't
generate an error because it was the last in the row to do so.

CHANGED: replaced all CMAKE_MODULE_PATH vars with KICAD_CMAKE_MODULE_DIR