SCH_BASE_FRAME: Implement SpaceMouse navigation and command export.
Added full support for using a 3Dconnexion device to pan & zoom in
Eeschema. Commands are exported and can be assigned to 3D mouse buttons.
The implementation is effectively a duplicate of that in PCB_BASE_FRAME
for pcbnew.exe.

Any limitations to the functionality are limitations of the installed
3Dconnexion driver for the device and OS.

Use build option KICAD_USE_3DCONNEXION (default = OFF) to control whether
the SpaceMouse support is compiled into the solution.