Fix SpaceMouse retains focus if the viewer window is iconized.
The window that has SpaceMouse focus is controlled by the wxActivate event.
wxWidgets does not send the wxActivateEvent when a window is minimized. On
Windows the window is iconized before the WM_ACTIVATE message is sent. The
fix adds handlers for the wxIconizeEvent to remove the focus from the
SpaceMouse target.