stm32_generator: Support additional NC pins
On some parts, DNU is used for what would otherwise be called NC pins.
In addition, some (power) pins are suffixed with "Unused". These pins
may still be connected to allow for compatibility with other parts in
the series (see section 5.3 of AN5031). These pins are left visible and
marked as passive to let the user decide what to do with them.