eeschema symbol editor derived symbols auto-rename and preserve footprint
Description of situation:
Making a duplicate of a symbol renames the VALUE field of the new
symbol to be the the same as the name of the old symbol if the VALUE
field of the old symbol was the same as the name of the old symbol.
Whereas making a derived symbol keeps the name of the parent symbol.

Additionally, a derived symbol has no footprint or datasheet copied from
the parent.

KiCAD library convention S6.2 says that the VALUE field of a symbol
should be the same as the name of the symbol.

The pinout of a derived symbol never differs from that of the parent
as the pinout is copied over upon derivation and all methods of changing
the pinout are disabled for the derived symbol.

Assertions:
As the pinout of the derived symbol cannot differ from the parent
usually the derived symbol will use the same footprint as the parent.
But not always, as often MSSOPs and SOICs will share pinouts and
SOT-23s and SC-70s will share pinouts.

Behavioral changes:
Change behavior so that derived symbol will have a VALUE field which is
the same as the name of the derived symbol if the VALUE field of the
parent symbol was the same as the name of the parent symbol.

Change behavior so that the derived symbol will copy the FOOTPRINT
field from the parent symbol.