Don't unwind collision in shove
The idea seems to be that we would remove a line  from the m_linestack
if one of the segments is our nearest collision.

The issue is that this removal can remove the line being routed, which
clears the tail and leaves the head dangling.

Fixes https://gitlab.com/kicad/code/kicad/-/issues/16591